Basic One-Seventh of night implementation
This implements the One-Seventh of night, as described in literature.

(refs: https://www.astronomycenter.net/pdf/2009_High_Latitude.pdf
https://www.astronomycenter.net/pdf/guessoum_2010.pdf)

TODO: Calculate or determine if we are in the abnormal period or not.
Currently, the method should be selected by the user manually. But it
would be nicer to provide an angle based alternative in the normal
period (when the astronomical signs of isha and fajr are visible.)