Add a way to customize the mode line message
In my Emacs installation, the unicode symbols used in mode line are not
displayed correctly. And to be honest, I didn't like the default format.
So I would like to see some thing like this in upstream, so I can
customize my mode line message.

The format is simple, it can contain an arbitrary text with ${hours},
${minutes}, and ${prayer}.